Unify the two files
