+++ src/st/cen/15714.c
@@ -108,3 +108,425 @@ void CollectLifeVessel(void) {
 }
 
 void DestroyCurrentEntity(void) { DestroyEntity(g_CurrentEntity); }
+
+u8 AnimateEntity(u8 frames[], Entity* entity);
+
+extern u16 D_80097408[];
+extern u8* D_80180F80[];
+extern s16 D_80180FE0[];
+extern s32 g_Status_subWeapon;
+
+void EntityPrizeDrop(Entity* self) {
+    Collider collider;
+    Primitive* prim;
+    s16 primIndex;
+    s16 var_a2;
+    u16 itemId;
+    s16 temp_a0;
+
+    itemId = self->params & 0x7FFF;
+    if (self->step != 0) {
+        AnimateEntity(D_80180F80[itemId], self);
+    }
+    if (self->step - 2 < 3U && self->hitFlags != 0) {
+        self->step = 5;
+    }
+    self->palette = 0;
+    if ((u8)self->unk6D[0] >= 0x18 && !(g_GameTimer & 2) && self->params != 1) {
+        self->palette = PAL_OVL(0x15F);
+    }
+    switch (self->step) {
+    case 0:
+        InitializeEntity(g_InitializeData0);
+        self->zPriority = g_unkGraphicsStruct.g_zEntityCenter.unk - 0x14;
+        self->drawMode = DRAW_DEFAULT;
+        if (itemId >= 0x18) {
+            self->params = 0;
+            itemId = 0;
+        }
+        if (itemId > 13 && itemId < 23 &&
+            itemId == D_80180F1C[g_Status.subWeapon]) {
+            self->params = itemId = 1;
+        }
+        if (itemId == 0 || itemId == 2) {
+            self->hitboxWidth = 4;
+        }
+        break;
+
+    case 1:
+        g_api.CheckCollision(self->posX.i.hi, self->posY.i.hi, &collider, 0);
+        if (collider.effects & EFFECT_NOTHROUGH_PLUS) {
+            DestroyEntity(self);
+        } else {
+            self->step++;
+            if (LOH(self->ext.generic.unk94) != 0) {
+                temp_a0 = LOH(self->ext.generic.unk94) - 1;
+                g_unkGraphicsStruct.g_zEntityCenter.S16 =
+                    g_unkGraphicsStruct.g_zEntityCenter.S16;
+                g_CastleFlags[(temp_a0 >> 3) + 0x1b0] |= 1 << (temp_a0 & 7);
+            }
+        }
+        if (itemId == 0) {
+            self->ext.generic.unk84.unk = -0x10000;
+            self->ext.generic.unk88.S16.unk0 = 0x800;
+        }
+        break;
+
+    case 2:
+        if (self->velocityY < 0) {
+            g_api.CheckCollision(
+                self->posX.i.hi, self->posY.i.hi - 7, &collider, 0);
+            if (collider.effects & EFFECT_NOTHROUGH) {
+                self->velocityY = 0;
+            }
+        }
+        MoveEntity();
+        g_api.CheckCollision(
+            self->posX.i.hi, self->posY.i.hi + 7, &collider, 0);
+        if (itemId != 0) {
+            if (collider.effects & EFFECT_NOTHROUGH && self->velocityY > 0) {
+                self->velocityX = 0;
+                self->velocityY = 0;
+                self->posY.i.hi += collider.unk18;
+                self->ext.generic.unk80.modeS8.unk0 = 0xF0;
+                self->step++;
+            } else {
+                FallEntity();
+            }
+            CheckFieldCollision(D_80180FE0, 2);
+        } else if (collider.effects & EFFECT_NOTHROUGH) {
+            self->posY.i.hi += collider.unk18;
+            self->ext.generic.unk80.modeS8.unk0 = 0x60;
+            self->step++;
+        } else {
+            func_80195714();
+        }
+        break;
+
+    case 3:
+        func_80195798(itemId);
+        if (!(self->params & 0x8000) &&
+            --self->ext.generic.unk80.modeS8.unk0 == 0) {
+            self->ext.generic.unk80.modeS8.unk0 = itemId == 0 ? 0x40 : 0x50;
+            self->step++;
+        }
+        break;
+
+    case 4:
+        func_80195798(itemId);
+        if (--self->ext.generic.unk80.modeS8.unk0) {
+            if (self->ext.generic.unk80.modeS8.unk0 & 2) {
+                self->animCurFrame = 0;
+            }
+        } else {
+            DestroyEntity(self);
+            return;
+        }
+        break;
+
+    case 5:
+        if (itemId == 0 || itemId == 1) {
+            CollectHeart(itemId);
+        } else if (itemId < 12) {
+            CollectGold(itemId);
+        } else if (itemId == 12) {
+            CollectHeartVessel();
+        } else if (itemId < 14) {
+            DestroyCurrentEntity();
+        } else if (itemId < 23) {
+            CollectSubweapon(itemId);
+        } else if (itemId == 23) {
+            CollectLifeVessel();
+        } else {
+            DestroyEntity(self);
+            return;
+        }
+        break;
+
+    case 6:
+    case 7:
+        switch (self->step_s) {
+        case 0:
+            self->animCurFrame = 0;
+            if (itemId > 13 && itemId < 23) {
+                if (itemId == D_80180F1C[g_Status.subWeapon]) {
+                    self->params = itemId = 1;
+                }
+            }
+            primIndex = g_api.AllocPrimitives(PRIM_GT4, 1);
+            if (primIndex != -1) {
+                self->primIndex = primIndex;
+                self->flags |= FLAG_HAS_PRIMS;
+                prim = &g_PrimBuf[primIndex];
+                prim->tpage = 0x1A;
+                prim->clut = 0x170;
+                prim->v2 = prim->v3 = 0x20;
+                prim->u1 = prim->u3 = 0x20;
+                prim->v0 = prim->v1 = 0;
+                prim->u0 = prim->u2 = 0;
+                prim->b0 = prim->b1 = prim->b2 = prim->b3 = 0x80;
+                prim->g0 = prim->g1 = prim->g2 = prim->g3 = 0x80;
+                prim->r0 = prim->r1 = prim->r2 = prim->r3 = 0x80;
+                prim->drawMode = DRAW_HIDE;
+                prim->priority = self->zPriority + 1;
+                self->step_s++;
+            }
+            break;
+
+        case 1:
+            MoveEntity();
+            g_api.CheckCollision(
+                self->posX.i.hi, self->posY.i.hi + 7, &collider, 0);
+            if (collider.effects & EFFECT_NOTHROUGH && self->velocityY > 0) {
+                self->velocityX = 0;
+                self->velocityY = 0;
+                self->posY.i.hi += collider.unk18;
+                self->step_s++;
+            } else {
+                FallEntity();
+            }
+            CheckFieldCollision(D_80180FE0, 2);
+            self->animCurFrame = 0;
+            if (self->ext.generic.unk88.S16.unk2 != 0) {
+                self->ext.generic.unk88.S16.unk2--;
+            } else {
+                prim = &g_PrimBuf[self->primIndex];
+                prim->x0 = prim->x2 = self->posX.i.hi - 1;
+                prim->x1 = prim->x3 = self->posX.i.hi + 1;
+                prim->y0 = prim->y1 = self->posY.i.hi - 1;
+                prim->y2 = prim->y3 = self->posY.i.hi + 1;
+                prim->drawMode = 0x37;
+            }
+            break;
+
+        case 2:
+            func_80195798(itemId);
+            prim = &g_PrimBuf[self->primIndex];
+            self->ext.generic.unk88.S16.unk2++;
+            if (self->ext.generic.unk88.S16.unk2 < 0x11) {
+                var_a2 = self->ext.generic.unk88.S16.unk2;
+                self->animCurFrame = 0;
+            } else {
+                var_a2 = 0x20 - self->ext.generic.unk88.S16.unk2;
+                prim->r0 = prim->r1 = prim->r2 = prim->r3 = prim->r3 - 8;
+                prim->g0 = prim->g1 = prim->g2 = prim->g3 = prim->g3 - 8;
+                prim->b0 = prim->b1 = prim->b2 = prim->b3 = prim->b3 - 8;
+            }
+            prim->x0 = prim->x2 = self->posX.i.hi - var_a2;
+            prim->x1 = prim->x3 = self->posX.i.hi + var_a2;
+            prim->y0 = prim->y1 = self->posY.i.hi - var_a2;
+            prim->y2 = prim->y3 = self->posY.i.hi + var_a2;
+            if (self->ext.generic.unk88.S16.unk2 == 0x20) {
+                g_api.FreePrimitives(self->primIndex);
+                self->ext.generic.unk80.modeS8.unk0 = 0xD0;
+                self->step = 3;
+                self->step_s = 0;
+                self->flags &= ~FLAG_HAS_PRIMS;
+            }
+            break;
+        }
+
+        break;
+    }
+}
+
+#define MISSING_ANIMATE_ENTITY_PROTOTYPE
+#include "../entity_explosion.h"
+#undef MISSING_ANIMATE_ENTITY_PROTOTYPE
+
+#include "../blink_item.h"
+
+static u16 g_ItemIconSlots[ICON_SLOT_NUM];
+
+extern s32 D_80097410;
+extern s32 D_80097414;
+void EntityEquipItemDrop(Entity* self) {
+    u16 itemId = self->params & 0x7FFF;
+    s32 primIndex;
+    Collider collider;
+    POLY_GT4* poly;
+    s32* itemName;
+    s16 index;
+    s32 vramX;
+    s32 vramY;
+    u8 left;
+    u8 right;
+    u16 top;
+    u8 bottom;
+    u8 temp_v0_6;
+    s16 temp_a0;
+    s32* unk;
+
+    if (((self->step - 2) < 3U) && (self->hitFlags != 0)) {
+        self->step = 5;
+    }
+
+    switch (self->step) {
+    case 0:
+        if (g_PlayableCharacter != PLAYER_ALUCARD) {
+            self->pfnUpdate = EntityPrizeDrop;
+            self->params = 0;
+            self->entityId = 3;
+            SetStep(0);
+            EntityPrizeDrop(self);
+            return;
+        }
+        InitializeEntity(g_InitializeData0);
+        self->ext.generic.unk7C.s = 0;
+        break;
+
+    case 1:
+        g_api.CheckCollision(self->posX.i.hi, self->posY.i.hi, &collider, 0);
+
+        if (!(collider.effects & EFFECT_NOTHROUGH_PLUS)) {
+            for (index = 0; index < 32; index++) {
+                if (g_ItemIconSlots[index] == 0) {
+                    break;
+                }
+            }
+
+            if (index >= 32) {
+                DestroyEntity(self);
+                return;
+            }
+
+            if (LOH(self->ext.generic.unk94) != 0) {
+                temp_a0 = LOH(self->ext.generic.unk94);
+                temp_a0--;
+                g_CastleFlags[(temp_a0 >> 3) + 0x1b0] |= 1 << (temp_a0 & 7);
+            }
+
+            primIndex = g_api.AllocPrimitives(PRIM_GT4, 1);
+            if (primIndex == -1) {
+                DestroyEntity(self);
+                return;
+            }
+
+            self->flags |= FLAG_HAS_PRIMS;
+            self->primIndex = primIndex;
+            g_ItemIconSlots[index] = 0x1E0;
+            self->ext.generic.unk8C.modeU16.unk0 = index;
+
+            if (itemId < NUM_HAND_ITEMS) {
+                g_api.LoadEquipIcon(g_api.equipDefs[itemId].icon,
+                                    g_api.equipDefs[itemId].iconPalette, index);
+            } else {
+                itemId -= NUM_HAND_ITEMS;
+                g_api.LoadEquipIcon(
+                    g_api.accessoryDefs[itemId].icon,
+                    g_api.accessoryDefs[itemId].iconPalette, index);
+            }
+
+            poly = &g_PrimBuf[primIndex];
+            vramX = ((temp_v0_6 = index) & 7) * 0x10;
+            vramY = (temp_v0_6 & 0x18) * 2;
+
+            left = vramX | 1;
+            right = vramX | 0xF;
+            top = 0x81;
+            top = vramY | top;
+            bottom = vramY | 0x8F;
+
+            poly->tpage = 0x1A;
+            poly->v3 = bottom;
+            poly->clut = index + 464;
+            poly->u0 = poly->u2 = left;
+            poly->u1 = poly->u3 = right;
+            poly->v1 = top;
+            poly->v2 = bottom;
+            poly->v0 = top;
+            poly->pad2 = 0x80;
+            poly->pad3 = 6;
+
+            self->ext.generic.unk7C.s = 128;
+            self->step++;
+            break;
+        }
+        DestroyEntity(self);
+        break;
+
+    case 2:
+        if (self->velocityY < 0) {
+            g_api.CheckCollision(
+                self->posX.i.hi, self->posY.i.hi - 7, &collider, 0);
+            if (collider.effects & EFFECT_NOTHROUGH) {
+                self->velocityY = 0;
+            }
+        }
+
+        MoveEntity();
+
+        g_api.CheckCollision(
+            self->posX.i.hi, self->posY.i.hi + 7, &collider, 0);
+
+        if ((collider.effects & EFFECT_NOTHROUGH) && (self->velocityY > 0)) {
+            self->velocityX = 0;
+            self->velocityY = 0;
+            self->posY.i.hi += collider.unk18;
+            self->ext.generic.unk80.modeS8.unk0 = 240;
+            self->step++;
+        } else {
+            FallEntity();
+        }
+
+        CheckFieldCollision(&D_80180FE0, 2);
+        break;
+
+    case 3:
+        func_80195798(1);
+        if (!(self->params & 0x8000)) {
+            if (!(--self->ext.generic.unk80.modeS8.unk0 & 255)) {
+                self->ext.generic.unk80.modeS8.unk0 = 0x50;
+                self->step++;
+            }
+        } else {
+            g_ItemIconSlots[self->ext.generic.unk8C.modeS16.unk0] = 0x10;
+        }
+        break;
+
+    case 4:
+        func_80195798(1);
+        if (self->ext.generic.unk80.modeS8.unk0 += 255) {
+            poly = &g_PrimBuf[self->primIndex];
+            if (self->ext.generic.unk80.modeS8.unk0 & 2) {
+                poly->pad3 = 8;
+            } else {
+                poly->pad3 = 2;
+            }
+        } else {
+            DestroyEntity(self);
+        }
+        break;
+
+    case 5:
+
+        unk = &D_80097410;
+        if (*unk) {
+            g_api.FreePrimitives(g_unkGraphicsStruct.BottomCornerTextPrims);
+            *unk = 0;
+        }
+
+        g_api.PlaySfx(SFX_ITEM_PICKUP);
+
+        if (itemId < NUM_HAND_ITEMS) {
+            itemName = g_api.equipDefs[itemId].name;
+            g_api.AddToInventory(itemId, EQUIP_HAND);
+        } else {
+            itemId -= NUM_HAND_ITEMS;
+            itemName = g_api.accessoryDefs[itemId].name;
+            g_api.AddToInventory(itemId, EQUIP_ARMOR);
+        }
+
+        BottomCornerText(itemName, 1);
+        DestroyEntity(self);
+        break;
+    }
+
+    if (self->step >= 2) {
+        if (self->ext.generic.unk7C.u != 0) {
+            self->ext.generic.unk7C.u--;
+        }
+        BlinkItem(self, self->ext.generic.unk7C.u);
+    }
+}
+
